requirements.txt
google-cloud-aiplatform>=1.38.0
python-dotenv>=1.0.0